--- bsp/stm32/stm32h743-atk-apollo/applications/drv_mpu.c
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2006-2018, RT-Thread Development Team
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- * Change Logs:
- * Date           Author       Notes
- * 2019-04-14     whj4674672   first version
- */
-#include <rtthread.h>
-#include "stm32h7xx.h"
-
-int mpu_init(void)
-{
-  MPU_Region_InitTypeDef MPU_InitStruct;
-  
-  /* Disable the MPU */
-  HAL_MPU_Disable();
-  
-  /* Configure the MPU attributes as WB for AXI SRAM */  
-  MPU_InitStruct.Enable		         = MPU_REGION_ENABLE;
-  MPU_InitStruct.BaseAddress	     = 0x24000000;
-  MPU_InitStruct.Size			         = MPU_REGION_SIZE_512KB;
-  MPU_InitStruct.AccessPermission  = MPU_REGION_FULL_ACCESS;
-  MPU_InitStruct.IsBufferable	     = MPU_ACCESS_BUFFERABLE;
-  MPU_InitStruct.IsCacheable 	     = MPU_ACCESS_CACHEABLE;
-  MPU_InitStruct.IsShareable	     = MPU_ACCESS_SHAREABLE;
-  MPU_InitStruct.Number		         = MPU_REGION_NUMBER0;
-  MPU_InitStruct.TypeExtField	     = MPU_TEX_LEVEL1;
-  MPU_InitStruct.SubRegionDisable  = 0X00;
-  MPU_InitStruct.DisableExec	     = MPU_INSTRUCTION_ACCESS_ENABLE;
-	
-  HAL_MPU_ConfigRegion(&MPU_InitStruct);
-
-#ifdef BSP_USING_SDRAM
-  /* Configure the MPU attributes as WT for SDRAM */
-  MPU_InitStruct.Enable            = MPU_REGION_ENABLE;
-  MPU_InitStruct.BaseAddress       = 0xC0000000;
-  MPU_InitStruct.Size              = MPU_REGION_SIZE_32MB;
-  MPU_InitStruct.AccessPermission  = MPU_REGION_FULL_ACCESS;
-  MPU_InitStruct.IsBufferable      = MPU_ACCESS_NOT_BUFFERABLE;
-  MPU_InitStruct.IsCacheable       = MPU_ACCESS_CACHEABLE;
-  MPU_InitStruct.IsShareable       = MPU_ACCESS_NOT_SHAREABLE;
-  MPU_InitStruct.Number            = MPU_REGION_NUMBER1;
-  MPU_InitStruct.TypeExtField      = MPU_TEX_LEVEL0;
-  MPU_InitStruct.SubRegionDisable  = 0x00;
-  MPU_InitStruct.DisableExec       = MPU_INSTRUCTION_ACCESS_ENABLE;
-
-  HAL_MPU_ConfigRegion(&MPU_InitStruct);
-#endif
-
-  /* Enable the MPU */
-  HAL_MPU_Enable(MPU_PRIVILEGED_DEFAULT);
-
-  return 0;
-  
-}
-INIT_BOARD_EXPORT(mpu_init);
